Move networking code to VirtusizeAPIService
